// apps/backend/src/generated/prisma/enums.ts
export const UserRole = {
  STUDENT: 'STUDENT',
  INSTRUCTOR: 'INSTRUCTOR',
  MANAGER: 'MANAGER',
  PRINCIPAL: 'PRINCIPAL',
  ADMIN: 'ADMIN',
} as const;

export type UserRole = (typeof UserRole)[keyof typeof UserRole];